ticket_api/requirements.txt
pandas
chromadb
google-genai
sentence-transformers
python-dotenv
fastapi
uvicorn[standard]
rank-bm25
